Keep service status validation in runtime
